Clean up of VTTests library
