+++ VTTests/RectMixins.cs
@@ -4,9 +4,7 @@ namespace VTTests
 {
     internal static class RectMixins
     {
-        public static Point Center(this Rect rect)
-        {
-            return new Point(rect.Left + rect.Width / 2, rect.Top + rect.Height / 2);
-        }
+        public static Point Center(this Rect rect) 
+            => new Point(rect.Left + rect.Width / 2, rect.Top + rect.Height / 2);
     }
 }
