--- VTTests/Test.cs
@@ -1,64 +1,0 @@
-﻿using System;
-using System.Diagnostics;
-using System.IO;
-using System.Reflection;
-using System.Windows;
-using GrpcDotNetNamedPipes;
-using VTTests.Internal;
-
-namespace VTTests
-{
-    public static class Test
-    {
-        internal const string PipePrefix = nameof(VTTest) + "ComminicationPipe";
-
-        public static IService StartService(Application? app = null)
-        {
-            var process = Process.GetCurrentProcess();
-            var service = new Service(process.Id.ToString(), app ?? Application.Current);
-            return service;
-        }
-
-        public static IApp ConnectToApp() 
-            => ConnectToApp(Process.GetCurrentProcess());
-
-        public static IApp ConnectToApp(int processId) 
-            => ConnectToApp(Process.GetProcessById(processId));
-
-        public static IApp ConnectToApp(Process process)
-        {
-            if (process is null)
-            {
-                throw new ArgumentNullException(nameof(process));
-            }
-
-            var channel = new NamedPipeChannel(".", PipePrefix + process.Id);
-            var client = new Protocol.ProtocolClient(channel);
-
-            return new App(client);
-        }
-
-        public static IApp StartApp(string? path = null)
-        {
-            path ??= Path.ChangeExtension(Assembly.GetExecutingAssembly().Location, ".exe");
-            path = Path.GetFullPath(path);
-            if (!File.Exists(path))
-            {
-                throw new Exception($"Could not find test app '{path}'");
-            }
-
-
-            var startInfo = new ProcessStartInfo(path)
-            {
-                WorkingDirectory = Path.GetDirectoryName(path) + Path.DirectorySeparatorChar,
-            };
-            Process process = Process.Start(startInfo);
-            var channel = new NamedPipeChannel(".", PipePrefix + process.Id);
-            var client = new Protocol.ProtocolClient(channel);
-
-            return new ManagedApp(process, client);
-        }
-
-
-    }
-}
