+++ VTTests/VTTests.csproj
@@ -9,13 +9,13 @@
   </PropertyGroup>
   
   <ItemGroup>
-    <PackageReference Include="Appium.WebDriver" Version="4.1.1" />
     <PackageReference Include="Grpc.Tools" Version="2.30.0">
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
     </PackageReference>
     <PackageReference Include="GrpcDotNetNamedPipes" Version="1.1.1" />
     <PackageReference Include="PInvoke.User32" Version="0.6.49" />
+    <PackageReference Include="System.Drawing.Common" Version="4.7.0" />
     
     <Protobuf Include="VTspec.proto" GrpcServices="Both" />
   </ItemGroup>
