+++ VTTests/Wait.cs
@@ -1,27 +1,9 @@
 ﻿using System;
 using System.Diagnostics;
-using System.Threading;
 using System.Threading.Tasks;
 
 namespace VTTests
 {
-    public class Retry
-    {
-        public int? MinAttempts { get; }
-        public TimeSpan Timeout { get; }
-
-        public Retry(int minAttempts, TimeSpan timeout)
-        {
-            MinAttempts = minAttempts;
-            Timeout = timeout;
-        }
-
-        public override string ToString() 
-            => $"{(MinAttempts != null ? $"Attempts: {MinAttempts}, " : "")}Timeout: {Timeout}";
-
-        public static Retry Default { get; } = new Retry(3, TimeSpan.FromSeconds(2));
-    }
-
     public static class Wait
     {
         public static async Task For(Func<Task<bool>> action, Retry? retry = null)
